rename file + add null check
